check_run.py added
Added scripts to check everything is running and acquiring properly on Pi. This is run at the start of each day through the witty_pi's afterStartup.sh script

+++ pycam/scripts/check_disk_space.py
@@ -7,51 +7,29 @@ deleted
 import subprocess
 import os
 import sys
+import datetime
 sys.path.append('/home/pi/')
 
 from pycam.setupclasses import FileLocator
 
-
-# Threshold (in MB)
-threshold = 100000   # 100 GB
-threshold_kb = threshold * 1000
-
-# Path to imag directory
+# Path to image directory
 img_path = FileLocator.IMG_SPEC_PATH
 
+del_days = [4, 8, 12, 16, 20, 24, 28]  # Days on which all existing data is deleted on local microSD
+date_now = datetime.datetime.now()
+day = date_now.day
 
-def get_storage_usage(img_path):
-    # TODO change from du to df, as du only gives file size, it doesn't quote total disk space - which is the important
-    # TODO thing
-    proc = subprocess.Popen(['du -s {}'.format(img_path)], stdout=subprocess.PIPE, shell=True)
-    stdout_value = proc.communicate()[0]
-    stdout_str = stdout_value.decode("utf-8")
-    stdout_lines = stdout_str.split('\n')
-
-    # Get kB storage by splitting the output string
-    storage_kb = int(stdout_lines[0].split()[0])
-    return storage_kb
-
-
-# Get storage
-storage_kb = get_storage_usage(img_path)
-print('Disk usage: {}'.format(storage_kb))
-print('Disk threshold: {}'.format(threshold_kb))
-
-while storage_kb > threshold_kb:
+if day in del_days:
     file_list = os.listdir(img_path)
-    file_list.sort()
-
-    # Catch exception just in case the file disappears before it can be removed
-    # (may get transferred then deleted by other program)
-    try:
-        file_path = os.path.join(img_path, file_list[0])
-        os.remove(file_path)
-        with open(FileLocator.REMOVED_FILES_LOG_PI, 'a') as f:
-            f.write('Data exceeding threshold disk space: {}MB. Removing file: {}\n'.format(threshold, file_path))
-    except BaseException as e:
-        with open(FileLocator.REMOVED_FILES_LOG_PI, 'a') as f:
-            f.write('{}\n'.format(e))
-
-    # Update storage
-    storage_kb = get_storage_usage(img_path)
+    for filename in file_list:
+        if len(file_list) < 10000:
+            break
+        # Catch exception just in case the file disappears before it can be removed
+        # (may get transferred then deleted by other program)
+        try:
+            file_path = os.path.join(img_path, filename)
+            os.remove(file_path)
+            print('Deleting file: {}'.format(filename))
+        except BaseException as e:
+            with open(FileLocator.REMOVED_FILES_LOG_PI, 'a') as f:
+                f.write('{}\n'.format(e))
